Removed some useless information.
Removed some useless information.

+++ examples/csrf_2.php
@@ -7,17 +7,17 @@ $security = new \security\CSRF;
 
 $token = $security->set(3, 3600);
 
-if(isset($_POST['token'])) {
+if($_SERVER['REQUEST_METHOD'] == 'POST') {
+ if(isset($_POST['token'])) {
   if($security->get($_POST['token'])) {
-    
-    $security->delete($_POST['token']); 
-    echo 'removed.';
+      
+   $security->delete($_POST['token']); 
+   echo 'removed.';
   }
+ }
 }
 ?>
-<div style="width: 400px;margin:0 auto;">
- <form action="#" method="post">
-  <input name="token" value="<?= $token; ?>">
-  <input type="submit">
- </form>
-</div>
+<form action="#" method="post">
+ <input name="token" value="<?= $token; ?>">
+ <input type="submit">
+</form>
